[Cli] Move Interaction Config class to Data namespace.
