+++ src/Valkyrja/Cli/Interaction/Data/Config.php
@@ -11,7 +11,7 @@ declare(strict_types=1);
  * file that was distributed with this source code.
  */
 
-namespace Valkyrja\Cli\Interaction;
+namespace Valkyrja\Cli\Interaction\Data;
 
 /**
  * Class Config.
