+++ src/Valkyrja/Cli/Interaction/Provider/ServiceProvider.php
@@ -15,7 +15,7 @@ namespace Valkyrja\Cli\Interaction\Provider;
 
 use Override;
 use Valkyrja\Application\Env;
-use Valkyrja\Cli\Interaction\Config;
+use Valkyrja\Cli\Interaction\Data\Config;
 use Valkyrja\Cli\Interaction\Factory\Contract\OutputFactory;
 use Valkyrja\Cli\Interaction\Factory\OutputFactory as DefaultOutputFactory;
 use Valkyrja\Container\Manager\Contract\Container;
